backend/requirements.txt
aiosmtplib==4.0.2
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.1
async-timeout==5.0.1
certifi==2025.8.3
cffi==2.0.0
click==8.1.8
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==4.2.23
django-allauth==65.10.0
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kombu==5.5.4
MarkupSafe==3.0.3
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.1.1
redis==6.2.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
vine==5.1.0
Werkzeug==3.1.3
